+++ Sources/ScrollKit/Examples/Examples+Spotify.swift
@@ -1,0 +1,15 @@
+//
+//  Examples+Spotify.swift
+//  ScrollKit
+//
+//  Created by Daniel Saidi on 2023-12-04.
+//  Copyright © 2023-2025 Daniel Saidi. All rights reserved.
+//
+
+import Foundation
+
+public extension Examples {
+ 
+    /// This namespace contains Spotify-related types.
+    struct Spotify {}
+}