+++ Sources/ScrollKit/Examples/Examples+SpotifyAlbumScreenHeader.swift
@@ -1,47 +1,47 @@
 //
-//  SpotifyPreviewScreenHeader.swift
+//  Examples+SpotifyAlbumScreenHeader.swift
 //  ScrollKit
 //
 //  Created by Daniel Saidi on 2023-02-06.
-//  Copyright © 2023-2024 Daniel Saidi. All rights reserved.
+//  Copyright © 2023-2025 Daniel Saidi. All rights reserved.
 //
 
 import SwiftUI
 
-public extension Spotify {
+public extension Examples.Spotify.AlbumScreen {
     
     /// This view mimics a Spotify album screen header.
-    struct PreviewScreenHeader: View {
+    struct Header: View {
         
         public init(
-            info: PreviewInfo,
-            headerVisibleRatio: CGFloat = 1
+            album: Examples.Spotify.Album,
+            visibleHeaderRatio: CGFloat = 1
         ) {
-            self.info = info
-            self.headerVisibleRatio = headerVisibleRatio
+            self.album = album
+            self.visibleHeaderRatio = visibleHeaderRatio
         }
         
         public static let height: CGFloat = 280
         
-        private var info: PreviewInfo
-        private var headerVisibleRatio: CGFloat
+        private var album: Examples.Spotify.Album
+        private var visibleHeaderRatio: CGFloat
         
         public var body: some View {
             ZStack {
-                ScrollViewHeaderGradient(info.tintColor, .black)
-                ScrollViewHeaderGradient(info.tintColor.opacity(1), info.tintColor.opacity(0))
-                    .opacity(1 - headerVisibleRatio)
+                ScrollViewHeaderGradient(album.tintColor, .black)
+                ScrollViewHeaderGradient(album.tintColor.opacity(1), album.tintColor.opacity(0))
+                    .opacity(1 - visibleHeaderRatio)
                 cover
             }
         }
     }
 }
 
-private extension Spotify.PreviewScreenHeader {
+private extension Examples.Spotify.AlbumScreen.Header {
 
     var cover: some View {
         AsyncImage(
-            url: URL(string: info.releaseCoverUrl),
+            url: URL(string: album.releaseCoverUrl),
             content: { image in
                 image.image?.resizable()
                     .aspectRatio(contentMode: .fit)
@@ -52,21 +52,21 @@ private extension Spotify.PreviewScreenHeader {
         .shadow(radius: 10)
         .rotation3DEffect(.degrees(rotationDegrees), axis: (x: 1, y: 0, z: 0))
         .offset(y: verticalOffset)
-        .opacity(headerVisibleRatio)
+        .opacity(visibleHeaderRatio)
         .padding(.top, 60)
         .padding(.bottom, 20)
         .padding(.horizontal, 20)
     }
 
     var rotationDegrees: CGFloat {
-        guard headerVisibleRatio > 1 else { return 0 }
-        let value = 20.0 * (1 - headerVisibleRatio)
+        guard visibleHeaderRatio > 1 else { return 0 }
+        let value = 20.0 * (1 - visibleHeaderRatio)
         return value.capped(to: -5...0)
     }
 
     var verticalOffset: CGFloat {
-        guard headerVisibleRatio < 1 else { return 0 }
-        return 70.0 * (1 - headerVisibleRatio)
+        guard visibleHeaderRatio < 1 else { return 0 }
+        return 70.0 * (1 - visibleHeaderRatio)
     }
 }
 
@@ -81,6 +81,6 @@ private extension CGFloat {
 
 #Preview {
 
-    Spotify.PreviewScreenHeader(info: .anthrax)
-        .frame(height: Spotify.PreviewScreenHeader.height)
+    Examples.Spotify.AlbumScreen.Header(album: .anthrax)
+        .frame(height: Examples.Spotify.AlbumScreen.Header.height)
 }