+++ Sources/ScrollKit/Examples/Examples.swift
@@ -1,0 +1,15 @@
+//
+//  Examples.swift
+//  ScrollKit
+//
+//  Created by Daniel Saidi on 2025-04-09.
+//  Copyright © 2025 Daniel Saidi. All rights reserved.
+//
+
+import Foundation
+
+/// This is a namespace with various examples on how you can
+/// use this library to achieve certain effects.
+///
+/// For now, only `
+public struct Examples {}